fix: some issues and also move stuff to skilltree package
