+++ src/common/main/kotlin/tech/thatgravyboat/skyblockapi/api/data/stored/HotmStorage.kt
@@ -4,7 +4,7 @@ import tech.thatgravyboat.skyblockapi.api.data.StoredProfileData
 import tech.thatgravyboat.skyblockapi.api.profile.hotm.HotmData
 import tech.thatgravyboat.skyblockapi.api.profile.hotm.HotmPerk
 
-internal object HotmStorage : HotxStorage<HotmData, HotmPerk>() {
+internal object HotmStorage : SkillTreeStorage<HotmData, HotmPerk>() {
 
     override var STORAGE = StoredProfileData(
         ::HotmData,
