+++ src/common/main/kotlin/tech/thatgravyboat/skyblockapi/api/data/stored/SkillTreeStorage.kt
@@ -1,10 +1,10 @@
 package tech.thatgravyboat.skyblockapi.api.data.stored
 
 import tech.thatgravyboat.skyblockapi.api.data.StoredProfileData
-import tech.thatgravyboat.skyblockapi.api.profile.hotx.SkillTreeData
-import tech.thatgravyboat.skyblockapi.api.profile.hotx.SkillTreePerk
+import tech.thatgravyboat.skyblockapi.api.profile.skilltree.SkillTreeData
+import tech.thatgravyboat.skyblockapi.api.profile.skilltree.SkillTreePerk
 
-internal abstract class HotxStorage<Data : SkillTreeData<Perk>, Perk : SkillTreePerk> {
+internal abstract class SkillTreeStorage<Data : SkillTreeData<Perk>, Perk : SkillTreePerk> {
 
     abstract val STORAGE: StoredProfileData<Data>
 
@@ -16,10 +16,10 @@ internal abstract class HotxStorage<Data : SkillTreeData<Perk>, Perk : SkillTree
         }
 
     var tokens: Int
-        get() = HotfStorage.STORAGE.get()?.tokens ?: 1
+        get() = STORAGE.get()?.tokens ?: 1
         internal set(value) {
             if (this.tokens == value) return
-            HotfStorage.STORAGE.get()?.tokens = value
+            STORAGE.get()?.tokens = value
             save()
         }
